+++ packages/sdk/src/codegen/templates/typescript/client.template.ts
@@ -1,3 +1,5 @@
+// TODO: Remap response type due to client mismatch
+
 //language=handlebars
 export const handlebarTemplate = `
 import {
@@ -13,13 +15,17 @@ import {
 	SubscriptionEventHandler,
 	FetchUserRequestOptions,
 	UploadValidationOptions,
+	QueryRequestOptions,
+  MutationRequestOptions,
+	ClientOperationErrors,
 	ExtractProfileName,
 	ExtractMeta,
+	GraphQLError
 } from "@wundergraph/sdk/client";
+import type { OperationErrors } from "./ts-operation-errors";
 
 import type { PublicCustomClaims } from "./claims";
 import type { {{ modelImports }} } from "./models";
-
 export type UserRole = {{{ roleDefinitions }}};
 
 export const WUNDERGRAPH_S3_ENABLED = {{hasS3Providers}};
@@ -106,26 +112,26 @@ export class WunderGraphClient extends Client {
 	query<
 		OperationName extends Extract<keyof Operations['queries'], string>,
 		Input extends Operations['queries'][OperationName]['input'] = Operations['queries'][OperationName]['input'],
-		Data extends Operations['queries'][OperationName]['data'] = Operations['queries'][OperationName]['data']
-	>(options: OperationName extends string ? OperationRequestOptions<OperationName, Input> : OperationRequestOptions) {
-		return super.query<OperationRequestOptions, Data>(options);
+    Response extends Operations['queries'][OperationName]['response'] = Operations['queries'][OperationName]['response']
+	>(options: OperationName extends string ? QueryRequestOptions<OperationName, Input> : OperationRequestOptions) {
+		return super.query<OperationRequestOptions, Response['data'], Response['error']>(options);
 	}
 	mutate<
 		OperationName extends Extract<keyof Operations['mutations'], string>,
 		Input extends Operations['mutations'][OperationName]['input'] = Operations['mutations'][OperationName]['input'],
-		Data extends Operations['mutations'][OperationName]['data'] = Operations['mutations'][OperationName]['data']
-	>(options: OperationName extends string ? OperationRequestOptions<OperationName, Input> : OperationRequestOptions) {
-		return super.mutate<OperationRequestOptions, Data>(options);
+    Response extends Operations['mutations'][OperationName]['response'] = Operations['mutations'][OperationName]['response']
+	>(options: OperationName extends string ? MutationRequestOptions<OperationName, Input> : OperationRequestOptions) {
+		return super.mutate<OperationRequestOptions, Response['data'], Response['error']>(options);
 	}
 	subscribe<
 		OperationName extends Extract<keyof Operations['subscriptions'], string>,
 		Input extends Operations['subscriptions'][OperationName]['input'] = Operations['subscriptions'][OperationName]['input'],
-		Data extends Operations['subscriptions'][OperationName]['data'] = Operations['subscriptions'][OperationName]['data']
+    Response extends Operations['subscriptions'][OperationName]['response'] = Operations['subscriptions'][OperationName]['response']
 	>(
 		options: OperationName extends string
 			? SubscriptionRequestOptions<OperationName, Input>
 			: SubscriptionRequestOptions,
-		cb: SubscriptionEventHandler<Data>
+		cb: SubscriptionEventHandler<Response['data'], Response['error']>
 	) {
 		return super.subscribe(options, cb);
 	}
@@ -165,7 +171,7 @@ export type Queries = {
 {{#each queries}}
     "{{operationPath}}": {
         {{#if hasInput}}input: {{operationName}}Input{{else}}input?: undefined{{/if}}
-        data: {{operationName}}ResponseData
+    		response: {{#if isTypeScriptOperation}}{ data?: {{operationName}}ResponseData, error?: OperationErrors['{{operationPath}}'] }{{else}}{ data?: {{operationName}}Response['data'], error?: ClientOperationErrors }{{/if}}
         requiresAuthentication: {{requiresAuthentication}}
         {{#if liveQuery}}liveQuery: boolean{{/if}}
     }
@@ -176,7 +182,7 @@ export type Mutations = {
 {{#each mutations}}
     "{{operationPath}}": {
         {{#if hasInput}}input: {{operationName}}Input{{else}}input?: undefined{{/if}}
-        data: {{operationName}}ResponseData
+    		response: {{#if isTypeScriptOperation}}{ data?: {{operationName}}ResponseData, error?: OperationErrors['{{operationPath}}'] }{{else}}{ data?: {{operationName}}Response['data'], error?: ClientOperationErrors }{{/if}}
         requiresAuthentication: {{requiresAuthentication}}
     }
 {{/each}}
@@ -186,7 +192,7 @@ export type Subscriptions = {
 {{#each subscriptions}}
     "{{operationPath}}": {
         {{#if hasInput}}input: {{operationName}}Input{{else}}input?: undefined{{/if}}
-        data: {{operationName}}ResponseData
+    		response: {{#if isTypeScriptOperation}}{ data?: {{operationName}}ResponseData, error?: OperationErrors['{{operationPath}}'] }{{else}}{ data?: {{operationName}}Response['data'], error?: ClientOperationErrors }{{/if}}
         requiresAuthentication: {{requiresAuthentication}}
     }
 {{/each}}
@@ -196,7 +202,7 @@ export type LiveQueries = {
 {{#each liveQueries}}
     "{{operationPath}}": {
         {{#if hasInput}}input: {{operationName}}Input{{else}}input?: undefined{{/if}}
-        data: {{operationName}}ResponseData
+    		response: {{#if isTypeScriptOperation}}{ data?: {{operationName}}ResponseData, error?: OperationErrors['{{operationPath}}'] }{{else}}{ data?: {{operationName}}Response['data'], error?: ClientOperationErrors }{{/if}}
         liveQuery: true
         requiresAuthentication: {{requiresAuthentication}}
     }