+++ testapps/default/test/index.test.ts
@@ -1,13 +1,11 @@
+import fs from 'fs/promises';
+import path from 'path';
+
 import { afterAll, beforeAll, describe, expect, test } from '@jest/globals';
 import { InputValidationError } from '@wundergraph/sdk/client';
 import { createTestServer } from '../.wundergraph/generated/testing';
 
 const wg = createTestServer();
-
-// TODO: Remove this once the secret refactor is done
-process.env['WG_SECURE_COOKIE_HASH_KEY'] = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
-process.env['WG_SECURE_COOKIE_BLOCK_KEY'] = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
-process.env['WG_CSRF_TOKEN_SECRET'] = 'aaaaaaaaaaa';
 
 beforeAll(() => wg.start());
 afterAll(() => wg.stop());
@@ -15,11 +13,21 @@ afterAll(() => wg.stop());
 describe('functions', () => {
 	test('internal operation call from  function', async () => {
 		const client = wg.client();
-		const result = await client.query({
-			operationName: 'nested/InternalWrapper',
-		});
-		expect(result.error).toBeUndefined();
-		expect(result.data?.data?.chinook_findFirstAlbum?.AlbumId).toBe(1);
+		const promises = [];
+		// Call this 50 times to exercise some code paths that share cached
+		// buffers. If we have a race condition in there, the go race detector
+		// (which we use in CI) will likely catch it.
+		for (let ii = 0; ii < 50; ii++) {
+			const op = client.query({
+				operationName: 'nested/InternalWrapper',
+			});
+			promises.push(op);
+		}
+		const results = await Promise.all(promises);
+		for (const result of results) {
+			expect(result.error).toBeUndefined();
+			expect(result.data?.data?.chinook_findFirstAlbum?.AlbumId).toBe(1);
+		}
 	});
 });
 
@@ -81,5 +89,109 @@ describe('@jsonSchema', () => {
 		let validationError = withoutCodeValidated.error as InputValidationError;
 		expect(validationError.errors[0].propertyPath).toBe('/filter/code/eq');
 		expect(validationError.errors[0].message).toMatch(/\[A-Z]\{2\}/);
+	});
+});
+
+describe('sql', () => {
+	test('execute raw', async () => {
+		const client = wg.client();
+		const result = await client.mutate({
+			operationName: 'rawsql/ExecuteRaw',
+			input: {
+				id: '2',
+				name: 'Jannik',
+				email: 'jannik@wundergraph.com',
+			},
+		});
+		expect(result.error).toBeUndefined();
+		expect(result.data?.users_post_executeRaw).toBe(0);
+		console.log(JSON.stringify(result));
+	});
+	test('execute raw inline', async () => {
+		const client = wg.client();
+		const result = await client.mutate({
+			operationName: 'rawsql/ExecuteRawInline',
+		});
+		expect(result.error).toBeUndefined();
+		expect(result.data?.users_post_executeRaw).toBe(0);
+		console.log(JSON.stringify(result));
+	});
+	test('query raw json', async () => {
+		const client = wg.client();
+		const result = await client.query({
+			operationName: 'rawsql/QueryRawJSON',
+		});
+		expect(result.error).toBeUndefined();
+		expect(JSON.stringify(result.data?.json)).toEqual(
+			`[{"id":1,"email":"jens@wundergraph.com","name":"Jens"},{"id":2,"email":"jannik@wundergraph.com","name":"Jannik"}]`
+		);
+	});
+	test('query raw row', async () => {
+		const client = wg.client();
+		const result = await client.query({
+			operationName: 'rawsql/QueryRow',
+			input: {
+				email: 'jens@wundergraph.com',
+			},
+		});
+		expect(result.error).toBeUndefined();
+		expect(result.data?.row[0].id).toEqual(1);
+		expect(result.data?.row[0].name).toEqual('Jens');
+		expect(result.data?.row[0].email).toEqual('jens@wundergraph.com');
+	});
+	test('query raw row inline', async () => {
+		const client = wg.client();
+		const result = await client.query({
+			operationName: 'rawsql/QueryRowInline',
+		});
+		expect(result.error).toBeUndefined();
+		console.log(JSON.stringify(result.data?.row));
+		expect(result.data?.row[0].id).toEqual(2);
+		expect(result.data?.row[0].name).toEqual('Jannik');
+		expect(result.data?.row[0].email).toEqual('jannik@wundergraph.com');
+	});
+});
+
+describe('@transform directive', () => {
+	test('@transform in public operation', async () => {
+		const result = await wg.client().query({
+			operationName: 'CountryWeather',
+			input: {
+				countryCode: 'ES',
+			},
+		});
+		expect(result.error).toBeUndefined();
+		const country = result?.data?.country?.[0];
+		expect(country).toBeDefined();
+		expect(country?.capital).toBe('Madrid');
+		expect(country?.weather.temperature?.max).toBeDefined();
+	});
+	test('@transform in internal operation', async () => {
+		const result = await wg.client().query({
+			operationName: 'CountryWeatherViaInternal',
+			input: {
+				countryCode: 'ES',
+			},
+		});
+		expect(result.error).toBeUndefined();
+		const country = result?.data?.data?.country?.[0];
+		expect(country).toBeDefined();
+		expect(country?.capital).toBe('Madrid');
+		expect(country?.weather.temperature?.max).toBeDefined();
+	});
+});
+
+describe('OpenAPI generation', () => {
+	test('OpenAPI includes TypeScript operation response schema', async () => {
+		const filePath = path.join(__dirname, '..', '.wundergraph', 'generated', 'wundergraph.openapi.json');
+		const data = await fs.readFile(filePath, { encoding: 'utf-8' });
+		const spec = JSON.parse(data);
+		const usersGet = spec.paths?.['/users/get'];
+		expect(usersGet).toBeDefined();
+		const response = usersGet?.['get']?.['responses']?.['200'];
+		expect(response).toBeDefined();
+		const responseSchema = response['content']?.['application/json']?.['schema'];
+		expect(responseSchema).toBeDefined();
+		expect(responseSchema['properties']?.['hello']?.['type']).toBe('string');
 	});
 });