+++ packages/swr/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.14.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.3...@wundergraph/swr@0.14.4) (2023-04-04)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.14.3](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.2...@wundergraph/swr@0.14.3) (2023-03-31)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.14.2](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.1...@wundergraph/swr@0.14.2) (2023-03-30)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.14.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.14.0...@wundergraph/swr@0.14.1) (2023-03-29)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.14.0](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.13.8...@wundergraph/swr@0.14.0) (2023-03-28)
+
+### Features
+
+* custom error support for typescript operations ([#759](https://github.com/wundergraph/wundergraph/issues/759)) ([7cfa815](https://github.com/wundergraph/wundergraph/commit/7cfa815a44b0b2a1184c2c50d0c4ea7cbb8066a7)) (@StarpTech)
+
+## [0.13.8](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.13.7...@wundergraph/swr@0.13.8) (2023-03-27)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.13.7](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.13.6...@wundergraph/swr@0.13.7) (2023-03-23)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.13.6](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.13.5...@wundergraph/swr@0.13.6) (2023-03-16)
+
+**Note:** Version bump only for package @wundergraph/swr
+
 ## [0.13.5](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.13.4...@wundergraph/swr@0.13.5) (2023-03-06)
 
 ### Bug Fixes